Oh boy, here we go
Imported custom theme generated in Theme roller.
Applied theme to V2 DeviceList and Device files for no good reason apperently.

Added Plus icon and click handler to Events listview in Device page so we can add custom events.  This to was in vain.  Will need to move this over to the main directory instead of V2.

Fuckin Trey.